requirements.txt
openai-whisper
google-generativeai
python-dotenv
ffmpeg-python
